Small bug fixes with respect to hex decoding.
git-svn-id: https://svn.apache.org/repos/asf/jakarta/poi/trunk@353012 13f79535-47bb-0310-9956-ffa450edef68

+++ src/contrib/src/org/apache/poi/contrib/poibrowser/PropertySetDescriptorRenderer.java
@@ -94,9 +94,9 @@ public class PropertySetDescriptorRenderer extends DocumentDescriptorRenderer
         text.setFont(new Font("Monospaced", Font.PLAIN, 10));
         text.append(renderAsString(d));
         text.append("\nByte order: " +
-                    Codec.hexEncode(ps.getByteOrder()));
+                    Codec.hexEncode((short) ps.getByteOrder()));
         text.append("\nFormat: " +
-                    Codec.hexEncode(ps.getFormat()));
+                    Codec.hexEncode((short) ps.getFormat()));
         text.append("\nOS version: " +
                     Codec.hexEncode(ps.getOSVersion()));
         text.append("\nClass ID: " +
